Fixed build errors

+++ src/object/cloud_particle_system.cpp
@@ -108,9 +108,9 @@ void CloudParticleSystem::update(float dt_sec)
     if (!cloudParticle)
       continue;
     cloudParticle->pos.x += cloudParticle->speed * dt_sec * m_current_speed;
-    if (cloudParticle->pos.x < cam.get_translation().x - cloudParticle->texture->get_width())
+    if (cloudParticle->pos.x < cam.get_translation().x - static_cast<float>(cloudParticle->texture->get_width()))
       cloudParticle->pos.x += virtual_width;
-    if (cloudParticle->pos.x > cam.get_translation().x + SCREEN_WIDTH)
+    if (cloudParticle->pos.x > cam.get_translation().x + static_cast<float>(SCREEN_WIDTH))
       cloudParticle->pos.x -= virtual_width;
 
     // Update alpha
